Added new arts to the Home Screen and the part are Top Rated and Popular where users can get info about those Animes.

+++ lib/customs/custom_methods.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class CustomMethods{
-  static double width(BuildContext context, int value) {
+  static double width(BuildContext context, double value) {
     return MediaQuery.of(context).size.width / value;
   }
 }
